Made the AI unbeatable
Rewrote buildGameTree_recursive as buildGameTree_recursive2 implementing
Mini Max Algorithm with Alpha Beta Pruning.

Adapted TreeNode to fit the new buildGameTree_recursive2.

Adapted GameGUI, AiWorker, MySwingWorker since signature of
buildGameTree_recursive2 differs from buildGameTree_recursive.

Changed scoreIfStateIsReached in GameJTable from int to double

+++ tic-tac-toe/TicTacToe/src/gui/GameJTable.java
@@ -21,7 +21,7 @@ public class GameJTable extends JTable {
 	 * Score assigned to this turn by the minimax-Algorithm of the AI while
 	 * computing all possible turns
 	 */
-	public int scoreIfStateIsReached;
+	public double scoreIfStateIsReached;
 
 	public GameJTable() {
 		super.setDefaultRenderer(String.class, new MyCellRenderer());
@@ -457,7 +457,7 @@ public class GameJTable extends JTable {
 	 * 
 	 * @return the negated scoreIfStateIsReached.
 	 */
-	public int invertScoreIfStateIsReached() {
+	public double invertScoreIfStateIsReached() {
 		scoreIfStateIsReached = -scoreIfStateIsReached;
 		return scoreIfStateIsReached;
 	}
